Move mock to dev
