+++ src/app/components/periodeliste/PeriodeHeader.tsx
@@ -8,11 +8,11 @@ import { InjectedIntlProps, injectIntl } from 'react-intl';
 import { Periode, Periodetype } from '../../types/uttaksplan/periodetyper';
 import { getPeriodeFarge } from '../../util/uttaksplan/styleUtils';
 import { Tidsperioden } from '../../util/uttaksplan/Tidsperioden';
-import UttaksplanIkon, { UttaksplanIkonKeys } from '../uttaksplanIkon/UttaksplanIkon';
+import UttaksplanIkon, { UttaksplanIkonKeys } from '../uttaksplan-ikon/UttaksplanIkon';
+import StønadskontoIkon from '../uttaksplan-ikon/St\u00F8nadskontoIkon';
+import UtsettelseIkon from '../uttaksplan-ikon/UtsettelseIkon';
 
 import './periodeheader.less';
-import StønadskontoIkon from '../uttaksplanIkon/St\u00F8nadskontoIkon';
-import UtsettelseIkon from '../uttaksplanIkon/UtsettelseIkon';
 
 export type AdvarselType = 'advarsel' | 'feil';
