+++ src/app/components/uttaksplan-ikon/StønadskontoIkon.tsx
