+++ src/app/components/uttaksplan-ikon/UtsettelseIkon.tsx
