+++ src/app/components/uttaksplan-ikon/UttaksplanIkon.tsx
