+++ src/app/components/uttaksplan-ikon/ikoner/AdvarselIkon.tsx
