+++ src/app/components/uttaksplan-ikon/ikoner/ArbeidIkon.tsx
