+++ src/app/components/uttaksplan-ikon/ikoner/FerieIkon.tsx
