+++ src/app/components/uttaksplan-ikon/ikoner/GradertIkon.tsx
