+++ src/app/components/uttaksplan-ikon/ikoner/PlasterIkon.tsx
