+++ src/app/components/uttaksplan-ikon/ikoner/SykehusIkon.tsx
