+++ src/app/components/uttaksplan-ikon/ikoner/TerminIkon.tsx
