+++ src/app/components/uttaksplan-ikon/ikoner/UttakIkon.tsx
